Extract expect method into utility class, edit resource tests
