+++ sentinel/src/main/java/gov/va/health/api/sentinel/SystemDefinitions.java
@@ -172,6 +172,7 @@ public class SystemDefinitions {
         .procedure("1200002649508")
         .procedures(
             Procedures.builder().fromDate("ge2009").onDate("ge2009").toDate("le2010").build())
+        .unknown("5555555555555")
         .build();
   }
 }
