+++ sentinel/src/main/java/gov/va/health/api/sentinel/TestIds.java
@@ -29,7 +29,7 @@ public class TestIds {
 
   @NonNull DiagnosticReports diagnosticReports;
   @NonNull Observations observations;
-  @NonNull PersonallyIdentifiableInformation pii;
+  PersonallyIdentifiableInformation pii;
   @NonNull Procedures procedures;
 
   @Value
