+++ sentinel/src/test/java/gov/va/health/api/sentinel/AllergyIntoleranceIT.java
@@ -6,13 +6,34 @@ import gov.va.api.health.argonaut.api.resources.AllergyIntolerance;
 import gov.va.api.health.argonaut.api.resources.OperationOutcome;
 import java.util.Arrays;
 import java.util.List;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
 
+@SuppressWarnings({"DefaultAnnotationParam", "WeakerAccess"})
+@RunWith(Parameterized.class)
+@Slf4j
 public class AllergyIntoleranceIT {
+
+  @Parameter(0)
+  public int status;
+
+  @Parameter(1)
+  public Class<?> response;
+
+  @Parameter(2)
+  public String path;
+
+  @Parameter(3)
+  public String[] params;
+
+  ResourceRequest resourceRequest = new ResourceRequest();
 
   @Parameters(name = "{index}: {0} {2}")
   public static List<Object[]> parameters() {
-    ResourceRequest resourceRequest = new ResourceRequest();
     TestIds ids = IdRegistrar.of(Sentinel.get().system()).registeredIds();
     return Arrays.asList(
         assertRequest(
@@ -36,5 +57,11 @@ public class AllergyIntoleranceIT {
             AllergyIntolerance.Bundle.class,
             "/api/AllergyIntolerance?patient={patient}",
             ids.patient()));
+  }
+
+  @Test
+  public void resourceRequestTest() {
+    resourceRequest.getResource(path, params, status, response);
+    resourceRequest.pagingParameterBounds(path, params, response);
   }
 }
