+++ sentinel/src/test/java/gov/va/health/api/sentinel/AppointmentIT.java
@@ -1,27 +1,54 @@
 package gov.va.health.api.sentinel;
 
+import static gov.va.health.api.sentinel.ResourceRequest.assertRequest;
+
 import gov.va.api.health.argonaut.api.resources.Appointment;
 import gov.va.api.health.argonaut.api.resources.OperationOutcome;
 import java.util.Arrays;
 import java.util.List;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
 
+@SuppressWarnings({"DefaultAnnotationParam", "WeakerAccess"})
+@RunWith(Parameterized.class)
+@Slf4j
 public class AppointmentIT {
+
+  @Parameter(0)
+  public int status;
+
+  @Parameter(1)
+  public Class<?> response;
+
+  @Parameter(2)
+  public String path;
+
+  @Parameter(3)
+  public String[] params;
+
+  ResourceRequest resourceRequest = new ResourceRequest();
 
   @Parameters(name = "{index}: {0} {2}")
   public static List<Object[]> parameters() {
-    ResourceRequest resourceRequest = new ResourceRequest();
     TestIds ids = IdRegistrar.of(Sentinel.get().system()).registeredIds();
     return Arrays.asList(
-        resourceRequest.assertRequest(
-            200, Appointment.class, "/api/Appointment/{id}", ids.appointment()),
-        resourceRequest.assertRequest(
-            404, OperationOutcome.class, "/api/Appointment/{id}", ids.unknown()),
-        resourceRequest.assertRequest(
+        assertRequest(200, Appointment.class, "/api/Appointment/{id}", ids.appointment()),
+        assertRequest(404, OperationOutcome.class, "/api/Appointment/{id}", ids.unknown()),
+        assertRequest(
             200, Appointment.Bundle.class, "/api/Appointment?_id={id}", ids.appointment()),
-        resourceRequest.assertRequest(
+        assertRequest(
             200, Appointment.Bundle.class, "/api/Appointment?identifier={id}", ids.appointment()),
-        resourceRequest.assertRequest(
+        assertRequest(
             200, Appointment.Bundle.class, "/api/Appointment?patient={patient}", ids.patient()));
+  }
+
+  @Test
+  public void resourceRequestTest() {
+    resourceRequest.getResource(path, params, status, response);
+    resourceRequest.pagingParameterBounds(path, params, response);
   }
 }
