+++ sentinel/src/test/java/gov/va/health/api/sentinel/ObservationAdvancedIT.java
@@ -5,13 +5,34 @@ import static gov.va.health.api.sentinel.ResourceRequest.assertRequest;
 import gov.va.api.health.argonaut.api.resources.Observation;
 import java.util.Arrays;
 import java.util.List;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameter;
 import org.junit.runners.Parameterized.Parameters;
 
+@SuppressWarnings({"DefaultAnnotationParam", "WeakerAccess"})
+@RunWith(Parameterized.class)
+@Slf4j
 public class ObservationAdvancedIT {
+
+  @Parameter(0)
+  public int status;
+
+  @Parameter(1)
+  public Class<?> response;
+
+  @Parameter(2)
+  public String path;
+
+  @Parameter(3)
+  public String[] params;
+
+  ResourceRequest resourceRequest = new ResourceRequest();
 
   @Parameters(name = "{index}: {0} {2}")
   public static List<Object[]> parameters() {
-    ResourceRequest resourceRequest = new ResourceRequest();
     TestIds ids = IdRegistrar.of(Sentinel.get().system()).registeredIds();
     return Arrays.asList(
         assertRequest(
@@ -55,5 +76,11 @@ public class ObservationAdvancedIT {
             ids.patient(),
             ids.observations().loinc1(),
             ids.observations().loinc2()));
+  }
+
+  @Test
+  public void resourceRequestTest() {
+    resourceRequest.getResource(path, params, status, response);
+    resourceRequest.pagingParameterBounds(path, params, response);
   }
 }
